FilterMicroservice/requirements.txt
paho-mqtt==2.1.0
pandas==2.2.1
nats-py==2.7.2